+++ src/multiband_gate.cpp
@@ -155,8 +155,10 @@ void MultibandGate::process(std::span<float>& left_in,
 
         // Normalize the current band gain reduction amount as a percentage,
         // where 0% is no gating, and 100% is a fully closed gate.
-        const float band_max_reduction_port_value = lv2_wrapper->get_control_port_value("gr_" + nstr);
-        // no reduction defaults to 1.0F; aka db_to_linear(0 dB)
+        // Double needed for the level bar widget.
+        const double band_max_reduction_port_value =
+            static_cast<double>(lv2_wrapper->get_control_port_value("gr_" + nstr));
+        // no reduction defaults to 1.0; aka db_to_linear(0 dB)
         gating_array.at(n) = util::normalize(reduction_port_array.at(n), band_max_reduction_port_value);
       }
